Move event types into optic-domain
