+++ workspaces/optic-domain/src/events.ts
@@ -1,0 +1,18 @@
+export enum ICoreShapeKinds {
+  ObjectKind = '$object',
+  ListKind = '$list',
+  MapKind = '$map',
+  OneOfKind = '$oneOf',
+  AnyKind = '$any',
+  StringKind = '$string',
+  NumberKind = '$number',
+  BooleanKind = '$boolean',
+  NullableKind = '$nullable',
+  OptionalKind = '$optional',
+  UnknownKind = '$unknown',
+}
+export enum ICoreShapeInnerParameterNames {
+  ListInner = '$listItem',
+  NullableInner = '$nullableInner',
+  OptionalInner = '$optionalInner',
+}
