+++ workspaces/optic-domain/src/index.ts
@@ -1,8 +1,10 @@
 export * as AsyncTools from './async-tools';
 export * as Streams from './streams';
 export * as Commands from './commands';
+export * as Events from './events';
 export * as Interactions from './interactions';
 
 export * from './commands';
+export * from './events';
 export * from './interactions';
 export * from './streams';
