+++ workspaces/ui-v2/src/lib/shape-diffs/field.ts
@@ -3,7 +3,6 @@ import { Actual, Expectation } from '../shape-diff-dsl-rust';
 import sortBy from 'lodash.sortby';
 import {
   CurrentSpecContext,
-  ICoreShapeKinds,
   IDiffDescription,
   IInteractionPreviewTab,
   IInterpretation,
@@ -17,6 +16,7 @@ import { builderInnerShapeFromChoices } from './build-inner-shape';
 import {
   AddField,
   FieldShapeFromShape,
+  ICoreShapeKinds,
   SetFieldShape,
   CQRSCommand,
 } from '@useoptic/optic-domain';
