+++ workspaces/ui-v2/src/lib/shape-diffs/list.ts
@@ -3,8 +3,6 @@ import { BodyShapeDiff } from '../parse-diff';
 import { Actual, Expectation } from '../shape-diff-dsl-rust';
 import {
   CurrentSpecContext,
-  ICoreShapeInnerParameterNames,
-  ICoreShapeKinds,
   IDiffDescription,
   IInteractionPreviewTab,
   IInterpretation,
@@ -13,6 +11,8 @@ import {
 import { code, plain } from '<src>/pages/diffs/components/ICopyRender';
 import { builderInnerShapeFromChoices } from './build-inner-shape';
 import {
+  ICoreShapeInnerParameterNames,
+  ICoreShapeKinds,
   ProviderInShape,
   SetParameterShape,
   ShapeProvider,
